feat: finalize 0.27 release hardening
