+++ .pre-commit-config.yaml
@@ -35,7 +35,7 @@ repos:
       - id: mixed-line-ending
       - id: trailing-whitespace
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: "v0.15.22"
+    rev: "v0.16.0"
     hooks:
       # Run the linter.
     - id: ruff
