+++ src/py/litestar_vite/plugin/_static.py
@@ -1,6 +1,15 @@
-"""Static files configuration dataclass."""
+"""Server-neutral static-serving contract.
+
+Describes where a Litestar Vite app's production static assets should be served
+from: either directly by the web server (``NATIVE``) or by Litestar's static
+router over ASGI (``ASGI``). The result is a plain dataclass with an explicit
+``placement`` discriminator so an external consumer such as litestar-granian can
+compare structurally (``config.placement == "native"``) without importing this
+package. This module deliberately carries no litestar-granian import.
+"""
 
 from dataclasses import dataclass
+from enum import Enum
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
@@ -19,6 +28,17 @@ if TYPE_CHECKING:
     )
 
 
+class StaticPlacement(str, Enum):
+    """Where static files are served from.
+
+    Subclassing ``str`` is deliberate: consumers compare structurally against the
+    literal value (``config.placement == "native"``) without importing this package.
+    """
+
+    NATIVE = "native"  # eligible: the web server may serve mounts directly
+    ASGI = "asgi"  # Litestar's static router must serve
+
+
 @dataclass(frozen=True, slots=True)
 class StaticServerMount:
     """Describe one static directory exposed to a native server."""
@@ -30,10 +50,34 @@ class StaticServerMount:
 
 @dataclass(frozen=True, slots=True)
 class StaticServerConfig:
-    """Describe native static mounts or why Litestar must serve them."""
+    """Describe where static files should be served from.
 
+    ``reason`` is diagnostic detail that is meaningful only when ``placement`` is
+    ``ASGI``; it explains why Litestar's static router must serve (development mode,
+    framework/SSR routing, manifest/build-state problems, and similar).
+    """
+
+    placement: StaticPlacement = StaticPlacement.ASGI
     mounts: tuple[StaticServerMount, ...] = ()
-    fallback_reason: str | None = None
+    reason: str | None = None
+
+    def __post_init__(self) -> None:
+        """Enforce the placement invariants.
+
+        Raises:
+            ValueError: If ``NATIVE`` has no mounts, ``NATIVE`` carries a reason, or
+                ``ASGI`` lacks a non-empty reason.
+        """
+        if self.placement is StaticPlacement.NATIVE:
+            if not self.mounts:
+                msg = "StaticServerConfig with NATIVE placement requires at least one mount."
+                raise ValueError(msg)
+            if self.reason is not None:
+                msg = "StaticServerConfig with NATIVE placement must not carry a reason; reason is ASGI-only detail."
+                raise ValueError(msg)
+        elif not self.reason:
+            msg = "StaticServerConfig with ASGI placement requires a non-empty reason."
+            raise ValueError(msg)
 
 
 @dataclass
